Cleaning reference databases
Some reference databases, like fish_db, have multiple peaks to a single cell, which pandas will not be able to cast as a float. This adds a function correct_value_type to attempt to address that, though I'm not sure what the scientifically appropriate thing to do would be in the case of multiple peaks.